MONOREPO: Update start.sh scripts to skip steps that are already done.

+++ shared/start-oe_showcase.sh
@@ -2,26 +2,31 @@
 
 set -ex
 
-# Stop containers in all packages.
-docker compose --project-directory=packages/oe_bootstrap_theme stop
-docker compose --project-directory=packages/oe_whitelabel stop
-docker compose --project-directory=packages/oe_showcase stop
-
 # Build assets in other packages.
-cd ../oe_bootstrap_theme
-docker compose up -d node
-docker compose exec -u node node npm install
-docker compose exec -u node node npm run build
-docker compose stop
+if [ ! -d ../oe_bootstrap_theme/assets/css ]; then
+  cd ../oe_bootstrap_theme
+  docker compose up -d node
+  docker compose exec -u node node npm install
+  docker compose exec -u node node npm run build
+  docker compose exec -u node rm -rf node_modules
+  cd ../oe_showcase
+fi
 
-cd ../oe_whitelabel
-docker compose up -d node
-docker compose exec -u node node npm install
-docker compose exec -u node node npm run build
-docker compose stop
+if [ ! -d ../oe_whitelabel/assets/css ]; then
+  cd ../oe_whitelabel
+  docker compose up -d node
+  docker compose exec -u node node npm install
+  docker compose exec -u node node npm run build
+  docker compose exec -u node rm -rf node_modules
+  cd ../oe_showcase
+fi
 
-cd ../oe_showcase
+# Stop containers in other packages.
+# This will also stop the node containers started earlier.
+docker compose --project-directory=../oe_bootstrap_theme stop
+docker compose --project-directory=../oe_whitelabel stop
 
+# Start containers in this package.
 docker compose up -d
 
 # Remove vendor and build directories for packages that will be symlinked.
@@ -37,15 +42,16 @@ docker compose exec web rm -rf ../../packages/oe_whitelabel/build/core
 docker compose exec web rm -rf ../../packages/oe_whitelabel/build/modules
 docker compose exec web rm -rf ../../packages/oe_whitelabel/build/themes
 
-cp composer.lock composer.mono.lock
-docker compose exec web composer update --no-install openeuropa/oe_whitelabel openeuropa/oe_bootstrap_theme
-
-docker compose exec web composer install
+if [ ! -f vendor/composer/installed.json ]; then
+  cp composer.lock composer.mono.lock
+  docker compose exec web composer update --no-install openeuropa/oe_whitelabel openeuropa/oe_bootstrap_theme
+  docker compose exec web composer install
+fi
 
 # See if Drupal is already installed.
-docker compose exec web ./vendor/bin/drush status | grep "DB name"
-
-if [ $? -eq 1 ]; then
+if docker compose exec web ./vendor/bin/drush status | egrep "^Database *: *Connected *$"; then
+  echo "Drupal is already installed."
+else
   # Drupal is not installed yet.
   docker compose exec web ./vendor/bin/run drupal:site-install
   docker-compose exec web ./vendor/bin/run ci:site-setup
